refactor(P-10): extract activity section into TaskActivitySection live component
Move comment form, activity timeline, actor display, and PubSub handling
out of TaskLive.Show into a self-contained live component.
